Fix set root from same boot device
In Clear Linux the root kernel parameter is fixed to
"PARTUUID=4f68bce3-e8cd-4db1-96e7-fbcaf984b709". This brings
an issue when boot from a USB device and the HD has the same
PartUUID: When booting from USB without initrd  the root
partition is mounted from the HD.

This patch adds a new bootflow: when the CLR PartUUID is
detected, set the root partition from the boot device.

Signed-off-by: Miguel Bernal Marin <miguel.bernal.marin@linux.intel.com>