Introducing linux-cpio sub-package
At 24e2b805aed5 ("add support for clr-init") was introduced a symlink to
a initrd file. Currently this method is deprecated and improved in
clr-init package.

Now as clr-init is used for crypted root partition, and when it is
used in a laptop without a USB keyboard, the user can not type the
password in. So we added a cpio file with i8042 module which will be
added as a initrd option to the bootloader. And the kernel will load both
files and merge clr-init and the cpio file to memory.