config: enable writeback throttling
I tested this feature with blogbench on my server and desktop + NVMe
SSDs. The server has 25% gain in the write test and 4% in the read test
with this feature enabled, while the desktop has 25% loss in the write
test(but still leading compared to the other distros we benchmarked) and
10% gain in the read test.

It's also enabled by default in many distros including Ubuntu 19.04,
RHEL 8, Debian Buster RC and Fedora 30. Can we enable it for Clear as well?

From the commit log:

This feature prevents the background writeback thread from stalling I/O,
e.g. reads.

For instance, without this option enabled, it's possible to stall reads
setting a writeback dirty threshold (/proc/sys/vm/dirty_writeback_centisecs)
sufficiently long.

Signed-off-by: Ming Chen <ming.a.chen@intel.com>