drop overload on wekeup patch
Dapeng1 Mi wrotes:

After running a thorough test, It looks 0131-overload-on-wakeup.patch is
a double-edged sword. Although this patch improves performance of some
workloads, such as sqlite, but it causes more regression on other workloads,
such as on stress-ng, hackbench, pgbench and rocksdb, etc,.
The whole Geometric mean of Clear Linux also declines to 1.09 from 1.11
since this patch.

Looking from the patch content, this patch seems too simple and arbitrary,
it could bypass the subsequent entire load balance logic. This looks not
reasonable. We may need to figure out a better patch for this (It could
take a long time). But before this, we may need to remove this patch form
Clear Kernel.