Enable binder and ashmem modules
Anbox is a free and open-source compatibility layer that aims to allow mobile applications and mobile games developed for Android to run on GNU/Linux distributions.
The application needs binder and ashmem modules to work. The project used to provide dkms modules alongside the application itself. The dkms modules, however, are not compatible with kernel versions 5.7.0+. So to be able to use anbox one needs to compile the modules which exist in kernel tree.
It's seems a good idea to compile the modules for 5.7.0+ kernels.