config: set DEBUG_FORCE_FUNCTION_ALIGN_32B
There are cases that a commit from one domain changes the function
address alignment of other domains, and cause magic performance bump
(regression or improvement).
Enable this option will help to verify if the bump is caused by function
alignment changes, while it will slightly increase the kernel size and
affect icache usage.

It is mainly for debug and performance tuning use.

resolves https://github.com/clearlinux/distribution/issues/2332