Add branch hints on most often hit branches in boot
Using gcov analysis of the kernel boot indicates that a handful of the
most heavily exercised branches can be optimized with same branch
hinting. Tested with 100 boots and we have ~0.25-0.30% improvement on
an Alderlake i9-12900.  Also exercised with microbenchmarks with
stress-ng to ensure we have no performance regressions on general
kernel performance.

Signed-off-by: Colin Ian King <colin.king@intel.com>